ivf_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sentence-transformers
faiss-cpu
numpy
supabase
groq
torch    # optional but recommended for sentence-transformers
